swagger adjustments and re-organize system endpoints
