--- apps/backend/src/health/controllers/health.controller.ts
@@ -1,15 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-import { HealthCheck, HealthCheckService } from '@nestjs/terminus';
-
-@ApiTags('Root')
-@Controller('health')
-export class HealthController {
-  constructor(private readonly health: HealthCheckService) {}
-
-  @Get()
-  @HealthCheck()
-  check() {
-    return this.health.check([]);
-  }
-}
