--- apps/backend/src/health/controllers/index.ts
@@ -1,1 +1,0 @@
-export { HealthController } from './health.controller';
