+++ apps/backend/src/system/dtos/get-version.response.ts
@@ -1,5 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Exclude, Expose } from 'class-transformer';
+
 @Exclude()
 export class GetVersionResponse {
   @ApiProperty({ example: '0.0.1' })
