+++ apps/backend/src/system/services/system.service.ts
@@ -1,5 +1,9 @@
-import { WebTMLogger } from '../../common/helpers/webtm-logger';
 import { Injectable } from '@nestjs/common';
+import { plainToInstance } from 'class-transformer';
+import { WebTMLogger } from '../../common/helpers/webtm-logger';
+import { getVersion } from '../../getVersion';
+import { GetVersionResponse } from '../dtos';
+
 @Injectable()
 export class SystemService {
   private readonly logger = new WebTMLogger(SystemService.name);
@@ -11,5 +15,10 @@ export class SystemService {
       image_processing_model: 'gpt-4o-mini',
     };
     return models;
+  }
+
+  getVersion(): GetVersionResponse {
+    const version = getVersion();
+    return plainToInstance(GetVersionResponse, { version });
   }
 }
