+++ apps/backend/src/system/system.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
+import { TerminusModule } from '@nestjs/terminus';
+import { CommonModule } from '../common/common.module';
 import { SystemController } from './controllers';
 import { SystemService } from './services';
-import { CommonModule } from '../common/common.module';
 
 @Module({
-  imports: [CommonModule],
+  imports: [CommonModule, TerminusModule],
   providers: [SystemService],
   controllers: [SystemController],
 })
